1. Added osme debug statements to server 2. Added canjoinchat convenience functions to chatroom management 3. Contacts view is compelte 4. Removed alertview category in favor of new subclass 5. Join chatroom from invite seems to work

+++ ios/chatter/chatter/ChatListViewController.m
@@ -322,7 +322,7 @@ const int MAX_RECENT_CHATS = 5;
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    int count = 0;
+    NSUInteger count = 0;
     if (section == 0) {
         count = [recentChatroomList count];
     }
